Add cookiecutter template
This commit introduces a 'cookiecutter' directory that users can use
to spin off new projects using exemplar as a template. After invoking
cookiecutter on the repository, users are prompted for a new project
name, directory name, C++ version, paper, owner, and description, and
these values are filled into the template.

It also adds two new checks into CI: that, when the template is filled
in with the values for exemplar itself, the result is identical to
exemplar; and, that if a name other than "exemplar" is specified for
the project name, then the string "exemplar" does not appear in the
output when the template is filled in.

These CI checks can be tested by maintainers of exemplar without
needing to update the pull request by running the script
check_cookiecutter.sh.