Fix duplicative CI runs when updating a pull request
When updating the branch associated with a pull request, previously,
CI would trigger twice: once for the update to the branch, and another
time for the update to the pull request itself.

This commit addresses the issue by restricting CI runs for updates to
branches to those that update the main branch itself.