Reduce verbosity of pull request template
This text contained a lot of guidance for authors opening pull
requests that amounted to general best practices that most developers
should already be aware of, which was annoying to have to remove every
time I opened a pull request.

Additionally, it contained guidance about our review process which was
outdated, and our review process is still in flux.

This commit strips it down to the bare essentials: our code of conduct
and the Beman Standard.