Update infra subdir
Problem
-------

These changes cannot be evaluated without corresponding changes to
the infra/ subdir.

Solution
--------

The corresponding changes were released to
github.com/beman-project/infra as a release candidate branch. Set
the SHA for the HEAD commit of that branch in the
`infra/.beman-submodule` file. Then run the `beman-submodule update`
command.