Remove note about omitting CMAKE_PREFIX_PATH
This states that if the user has already installed a packaged version
of `infra/cmake/beman-install-library-config.cmake`, they can omit
`-DCMAKE_PREFIX_PATH=./infra/cmake`. However, no such packaged version
exists; the only way to use it currently is through the vendored copy,
so this note is not useful in practice. We can restore it when
`beman-install-library-config` is packaged.