WIP Draft: Convert exemplar into an INTERFACE library
This commit eliminates the src/ directory, moving
beman.examplar-config.cmake.in into the top level directory, moving
the contents of src/beman/exemplar/CMakeLists.txt into the top level
CMakeLists.txt, removing identity.cpp, and changing beman.exemplar
from STATIC/SHARED into INTERFACE.

This commit is currently just intended to be a jumping-off point for
discussion. What I would like to do if we are okay with this direction
is to preserve the option for users to create STATIC/SHARED libraries
by configuring cookiecutter differently. That feature is not present
yet here.